Move FileProcessListenerExtensions to detekt-test
