+++ detekt-metrics/build.gradle.kts
@@ -6,6 +6,7 @@ dependencies {
     api(projects.detektApi)
     api(libs.kotlin.compiler)
     compileOnly(projects.detektPsiUtils)
+    testImplementation(projects.detektTest)
     testImplementation(projects.detektTestUtils)
     testImplementation(testFixtures(projects.detektApi))
     testImplementation(libs.assertj.core)
