+++ detekt-metrics/src/test/kotlin/dev/detekt/metrics/processors/ProjectCognitiveComplexityProcessorSpec.kt
@@ -1,6 +1,7 @@
 package dev.detekt.metrics.processors
 
 import dev.detekt.metrics.CognitiveComplexity
+import dev.detekt.test.invoke
 import dev.detekt.test.utils.compileContentForTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
