+++ detekt-test/build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 
 dependencies {
     api(projects.detektApi)
+    implementation(testFixtures(projects.detektApi))
     api(projects.detektTestUtils)
     api(libs.kotlin.compiler)
     implementation(libs.kotlin.reflect)
